backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
boto3==1.34.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
Pillow==10.2.0
pytesseract==0.3.10
